Doc file to be submitted
Here's the file with all the input and outputs along with the snapshots of the test cases. Edit the credentials and submit it in the google form. And also complete the coding the Hackerrank too.
Hackerank Assignment link : http://www.hackerrank.com/ds-algo-lab-week-5
Assignment Submission link :  https://forms.gle/Nb6NuvzxpiyfhGNN6